[refactor] - Numeric improvements:
* Simplify pointless arithmetic expression.
* Integer multiplication  cast to long.
* Use valueOf
* Simplify pointless arithmetic expression

+++ mission/simulator/opssat-spacecraft-simulator/src/main/java/opssat/simulator/util/wav/WavFile.java
@@ -396,7 +396,7 @@ public class WavFile
 
 			int v = buffer[bufferPointer];
 			if (b < bytesPerSample-1 || bytesPerSample == 1) v &= 0xFF;
-			val += v << (b * 8);
+			val += (long) v << (b * 8);
 
 			bufferPointer ++;
 		}
